web_interface/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.41.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
alembic==1.13.1
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.3.0
python-nmap==0.7.1
paramiko==3.4.0
scapy==2.5.0
cryptography==41.0.7
jinja2==3.1.2
markdown==3.5.1
python-dateutil==2.8.2
pytz==2023.3.post1
colorama==0.4.6
termcolor==2.3.0
pygments==2.17.2